disable old cinder scheduler service
'ceph' was the cinder scheduler service host name. As we add v7k in this
release. So we use new host name for cinder scheduler service. After
upgrade, we can see new scheduler services are up, old service is down.
sensu alert error because of service down. 
This commit is to disable old scheduler service.