Allow for custom m12 configuration
The current jinja2 template for m12_plugin.ini doesn't allow for extra configurations to be added or for configurations to be moved. For example, if someone wants to not use the vlan tunnel type and configuraton, they don't have the ability to turn it off. This would allow users to overide the m12_plugin jinja2 directly in the ansible inventory for more specialized deployments.